+++ lib/ratistics/collection.rb
@@ -3,6 +3,65 @@ module Ratistics
   module Collection
     extend self
 
+    def binary_search(data, key, opts={}, &block)
+      return nil if data.nil? || data.empty?
+      data = data.sort unless block_given? || opts[:sorted] == true
+
+      imin = [opts[:imin].to_i, 0].max
+      imax = opts[:imax].nil? ? data.size-1 : [opts[:imax], data.size-1].min
+
+      if block_given?
+        min, max = yield(data[imin]), yield(data[imax])
+      else
+        min, max = data[imin], data[imax]
+      end
+      return [nil, imin] if key < min
+      return [imin, imin] if key == min
+      return [imax, nil] if key > max
+      return [imax, imax] if key == max
+
+      while (imax >= imin)
+        imid = (imin + imax) / 2
+        current = data[imid]
+        current = yield(current) if block_given?
+        if current < key
+          imin = imid + 1
+        elsif current > key
+          imax = imid - 1
+        else
+          imin = imax = imid
+          break
+        end
+      end
+
+      return imax, imin
+    end
+
+    alias :bsearch :binary_search
+    alias :half_interval_search :binary_search
+
+    def ascending?(data, opts={}, &block)
+      (data.size-1).times do |i|
+        if block_given?
+          return false if yield(data[i]) > yield(data[i+1])
+        else
+          return false if data[i] > data[i+1]
+        end
+      end
+      return true
+    end
+
+    def descending?(data, opts={}, &block)
+      (data.size-1).times do |i|
+        if block_given?
+          return false if yield(data[i]) < yield(data[i+1])
+        else
+          return false if data[i] < data[i+1]
+        end
+      end
+      return true
+    end
+
     # Override of #slice from Ruby Array. Provides a consistent interface
     # to slice data structures that do not have a native #slice method.
     #